铁路货运 毕设 （ 预完成）

+++ service/src/main/java/com/hrbu/service/organized/OrganizeServiceImpl.java
@@ -1,17 +1,27 @@
 package com.hrbu.service.organized;
 
 import com.hrbu.domain.Organize;
+import com.hrbu.domain.Province;
+import com.hrbu.mapper.orgProvince.OrgProvinceMapper;
 import com.hrbu.mapper.organize.OrganizeMapper;
+import com.hrbu.mapper.provine.ProvinceMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import java.util.*;
+
 @Service
 public class OrganizeServiceImpl implements OrganizeService {
 
 
     @Autowired
     private OrganizeMapper organizeMapper;
+
+    @Autowired
+    private OrgProvinceMapper orgProvinceMapper;
+
+    @Autowired
+    private ProvinceMapper provinceMapper;
 
     @Override
     public int selectCount() throws Exception{
@@ -33,16 +43,30 @@ public class OrganizeServiceImpl implements OrganizeService {
     }
 
     @Override
-    public boolean saveOrg(Organize organize) throws Exception{
-
+    public boolean saveOrg(Organize organize,String[] proId) throws Exception{
+        organize.setOrgId(UUID.randomUUID().toString());
+        organize.setCreateTime(new Date());
         int count = organizeMapper.saveOrg(organize);
-        return count>0;
+        //在这里将机构的管辖区域 添加进来
+        if(count>0){
+            Map map = new HashMap();
+            map.put("proId",proId);
+            map.put("orgId",organize.getOrgId());
+            int count2 = orgProvinceMapper.insertProvinceOrganize(map);
+            return count2>0;
+        }
+        return false;
     }
 
     @Override
-    public void deleteOrg(String orgId) throws Exception{
-
-        organizeMapper.deleteOrg(orgId);
+    public boolean deleteOrg(String orgId) throws Exception{
+//删除机构 要先删除关联表 t_org_pro  t_train
+        int preDel = orgProvinceMapper.deleteProOrg(orgId);
+        if (preDel>0){
+            int count = organizeMapper.deleteOrg(orgId);
+            return count>0;
+        }
+        return false;
     }
 
     @Override
@@ -51,11 +75,30 @@ public class OrganizeServiceImpl implements OrganizeService {
         Organize organize = organizeMapper.findOrgById(orgId);
         return organize;
     }
+//找到proId 界面上显示 管辖
+    @Override
+    public List<Province> toUpdateSelectproId(String orgId) throws Exception {
+        List<Province> proList = orgProvinceMapper.selectProIdByorgId(orgId);
+        return proList;
+    }
 
     @Override
-    public void updateOrg(Organize organize) throws Exception{
+    public boolean updateOrg(Organize organize,String[] proId) throws Exception{
 
-        organizeMapper.updateOrg(organize);
+        int count = organizeMapper.updateOrg(organize);
+        //修改 机构的管辖区域  由于是一对多 实现修改要先删(关系表)后添
+        if (count>0){
+
+            int delCount = orgProvinceMapper.deleteProOrg(organize.getOrgId());//shan
+            //tian
+                Map map = new HashMap();
+                map.put("proId",proId);
+                map.put("orgId",organize.getOrgId());
+                int addCount = orgProvinceMapper.insertProvinceOrganize(map);
+                return addCount>0;
+        }
+
+        return false;
     }
 
     @Override
